Remove pointless setting of autocommit status on connection that is only used to setup and destroy the test state.
Move the closing of xaconn up in the tearDown method to avoid a
deadlock on some test failures.  The xaconn has touched the testxa1
table which takes out a share lock on it.  If for some reason the test
fails to commit, the tearDown method tries to drop the table but
cannot it because there is still an outstanding lock on it.